I HATE CIRCULAR DEPENDENCIES

+++ include/neuron.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <stdint.h>
 #include <vector>
 
 namespace SirHoog
@@ -11,13 +10,13 @@ namespace SirHoog
             float activation = 0;
             std::vector<float> weights = {};
 
-            Neuron(float activation = 0, std::vector<float> weights = {}, uint8_t weightSize = 0)
+            Neuron(float activation = 0, std::vector<float> weights = {}, int weightSize = 0)
             {
                 this->activation = activation;
 
                 if (weightSize > 0)
                 {
-                    for (uint8_t i = 0; i < weightSize; i++)
+                    for (int i = 0; i < weightSize; i++)
                     {
                         weights.push_back(0.f);
                     }
